Move interceptor tests to separate test class
