+++ tests/NetEscapades.EnumGenerators.Tests/EnumGeneratorTests.cs
@@ -15,9 +15,6 @@ namespace NetEscapades.EnumGenerators.Tests.Roslyn4_04;
 [UsesVerify]
 public class EnumGeneratorTests
 {
-    private readonly Dictionary<string, string> _interceptionEnabled =
-        new() { { "build_property.EnableEnumGeneratorInterceptor", "true" } };
-
     [Fact]
     public Task CanGenerateEnumExtensionsInGlobalNamespace()
     {
@@ -343,309 +340,6 @@ namespace Foo
         Assert.Empty(diagnostics);
         return Verifier.Verify(output).ScrubExpectedChanges().UseDirectory("Snapshots");
     }
-
-#if INTERCEPTORS
-    [Fact]
-    public Task CanInterceptEnum()
-    {
-        const string input =
-            """
-            using NetEscapades.EnumGenerators;
-
-            namespace MyTestNameSpace
-            {
-                [EnumExtensions]
-                internal enum MyEnum
-                {
-                    First = 0,
-                    Second = 1,
-                }
-                
-                public class InnerClass
-                {
-                    public MyEnum _field = default;
-                    public MyEnum Property {get;set;} = default;
-                    public void MyTest()
-                    {
-                        var myValue = MyEnum.Second;
-                        var var1 = myValue.ToString();
-                        var var2 = MyEnum.Second.ToString();
-                        var var3 = Property.ToString();
-                        var var4 = _field.ToString();
-                    }
-                }
-            }
-            """;
-        var (diagnostics, output) =
-            TestHelpers.GetGeneratedTrees<EnumGenerator, TrackingNames>(new(_interceptionEnabled, input));
-
-        Assert.Empty(diagnostics);
-        return Verifier.Verify(output).ScrubExpectedChanges().UseDirectory("Snapshots");
-    }
-
-    [Fact]
-    public Task CanInterceptEnum2()
-    {
-        const string input =
-            """
-            using NetEscapades.EnumGenerators;
-
-            namespace MyTestNameSpace
-            {
-                [EnumExtensions]
-                internal enum MyEnum
-                {
-                    First = 0,
-                    Second = 1,
-                    Third = 2,
-                }
-
-                [EnumExtensions]
-                internal enum AnotherEnum
-                {
-                    First = 0,
-                    Second = 1,
-                    Third = 2,
-                }
-
-                internal enum YetAnotherEnum
-                {
-                    First = 0,
-                    Second = 1,
-                    Third = 2,
-                }
-                
-                public class InnerClass
-                {
-                    public void MyTest()
-                    {
-                        var result = AnotherEnum.First.ToString();
-                        AssertValue(MyEnum.First);
-                        AssertValue(AnotherEnum.Second);
-                        AssertValue(MyEnum.Third);
-                        var result2 = YetAnotherEnum.First.ToString();
-                        AssertValue(YetAnotherEnum.Second);
-                        
-                        void AssertValue(MyEnum value)
-                        {
-                            var toString = value.ToString();
-                        }
-
-                        void AssertValue(AnotherEnum value)
-                        {
-                            var toString = value.ToString();
-                        }
-
-                        void AssertValue(YetAnotherEnum value)
-                        {
-                            var toString = value.ToString();
-                        }
-                    }
-                }
-            }
-            """;
-        var (diagnostics, output) = 
-            TestHelpers.GetGeneratedTrees<EnumGenerator, TrackingNames>(new(_interceptionEnabled, input));
-
-        Assert.Empty(diagnostics);
-        return Verifier.Verify(output).ScrubExpectedChanges().UseDirectory("Snapshots");
-    }
-
-    [Fact]
-    public Task CanInterceptExternalEnum()
-    {
-        const string input =
-            """
-            using System;
-            using NetEscapades.EnumGenerators;
-            
-            [assembly:EnumExtensions<StringComparison>()]
-
-            namespace MyTestNameSpace
-            {
-                public class InnerClass
-                {
-                    public StringComparison _field = default;
-                    public StringComparison Property {get;set;} = default;
-                    public void MyTest()
-                    {
-                        var myValue = StringComparison.Ordinal;
-                        var var1 = myValue.ToString();
-                        var var2 = StringComparison.Ordinal.ToString();
-                        var var3 = Property.ToString();
-                        var var4 = _field.ToString();
-                        AssertValue(StringComparison.OrdinalIgnoreCase);
-
-                        void AssertValue(StringComparison value)
-                        {
-                            var toString = value.ToString();
-                        }
-                    }
-                }
-            }
-            """;
-        var (diagnostics, output) = 
-            TestHelpers.GetGeneratedTrees<EnumGenerator, TrackingNames>(new(_interceptionEnabled, input));
-
-        Assert.Empty(diagnostics);
-        return Verifier.Verify(output).ScrubExpectedChanges().UseDirectory("Snapshots");
-    }
-
-    [Fact]
-    public Task DoesNotInterceptWhenDisabled()
-    {
-        const string input =
-            """
-            using NetEscapades.EnumGenerators;
-
-            namespace MyTestNameSpace
-            {
-                [EnumExtensions]
-                internal enum MyEnum
-                {
-                    First = 0,
-                    Second = 1,
-                }
-                
-                public class InnerClass
-                {
-                    public MyEnum _field = default;
-                    public MyEnum Property {get;set;} = default;
-                    public void MyTest()
-                    {
-                        var myValue = MyEnum.Second;
-                        var var1 = myValue.ToString();
-                        var var2 = MyEnum.Second.ToString();
-                        var var3 = Property.ToString();
-                        var var4 = _field.ToString();
-                    }
-                }
-            }
-            """;
-        var (diagnostics, output) =
-            TestHelpers.GetGeneratedTrees<EnumGenerator, TrackingNames>(new(input));
-
-        Assert.Empty(diagnostics);
-        return Verifier.Verify(output).ScrubExpectedChanges().UseDirectory("Snapshots");
-    }
-
-    [Fact]
-    public Task DoesNotInterceptWhenOldCsharp()
-    {
-        const string input =
-            """
-            using NetEscapades.EnumGenerators;
-
-            namespace MyTestNameSpace
-            {
-                [EnumExtensions]
-                internal enum MyEnum
-                {
-                    First = 0,
-                    Second = 1,
-                }
-                
-                public class InnerClass
-                {
-                    public MyEnum _field = default;
-                    public MyEnum Property {get;set;} = default;
-                    public void MyTest()
-                    {
-                        var myValue = MyEnum.Second;
-                        var var1 = myValue.ToString();
-                        var var2 = MyEnum.Second.ToString();
-                        var var3 = Property.ToString();
-                        var var4 = _field.ToString();
-                    }
-                }
-            }
-            """;
-        var opts = new TestHelpers.Options(LanguageVersion.CSharp10, _interceptionEnabled, input);
-        var (diagnostics, output) =
-            TestHelpers.GetGeneratedTrees<EnumGenerator, TrackingNames>(opts);
-
-        diagnostics.Should().ContainSingle(x => x.Id == DiagnosticHelper.CsharpVersionLooLow.Id);
-        return Verifier.Verify(output).ScrubExpectedChanges().UseDirectory("Snapshots");
-    }
-
-    [Fact]
-    public Task CanInterceptWhenCsharp11()
-    {
-        const string input =
-            """
-            using NetEscapades.EnumGenerators;
-
-            namespace MyTestNameSpace
-            {
-                [EnumExtensions]
-                internal enum MyEnum
-                {
-                    First = 0,
-                    Second = 1,
-                }
-                
-                public class InnerClass
-                {
-                    public MyEnum _field = default;
-                    public MyEnum Property {get;set;} = default;
-                    public void MyTest()
-                    {
-                        var myValue = MyEnum.Second;
-                        var var1 = myValue.ToString();
-                        var var2 = MyEnum.Second.ToString();
-                        var var3 = Property.ToString();
-                        var var4 = _field.ToString();
-                    }
-                }
-            }
-            """;
-        var opts = new TestHelpers.Options(LanguageVersion.CSharp11, _interceptionEnabled, input);
-        var (diagnostics, output) =
-            TestHelpers.GetGeneratedTrees<EnumGenerator, TrackingNames>(opts);
-
-        Assert.Empty(diagnostics);
-        return Verifier.Verify(output).ScrubExpectedChanges().UseDirectory("Snapshots");
-    }
-#else
-    [Fact]
-    public Task CanNotInterceptEnum()
-    {
-        const string input =
-            """
-            using NetEscapades.EnumGenerators;
-
-            namespace MyTestNameSpace
-            {
-                [EnumExtensions]
-                internal enum MyEnum
-                {
-                    First = 0,
-                    Second = 1,
-                }
-                
-                public class InnerClass
-                {
-                    public MyEnum _field = default;
-                    public MyEnum Property {get;set;} = default;
-                    public void MyTest()
-                    {
-                        var myValue = MyEnum.Second;
-                        var var1 = myValue.ToString();
-                        var var2 = MyEnum.Second.ToString();
-                        var var3 = Property.ToString();
-                        var var4 = _field.ToString();
-                    }
-                }
-            }
-            """;
-        var (diagnostics, output) =
-            TestHelpers.GetGeneratedTrees<EnumGenerator, TrackingNames>(new(_interceptionEnabled, input));
-
-        diagnostics.Should().ContainSingle(x => x.Id == DiagnosticHelper.SdkVersionTooLow.Id);
-        return Verifier.Verify(output).ScrubExpectedChanges().UseDirectory("Snapshots");
-    }
-
-#endif
 
     [Fact]
     public Task CanGenerateForExternalEnum()
